django_backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
stripe==7.4.0
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.1
pymongo==4.6.0
